make toolbar component
simplify demos
