--- sites/demo/src/containers/FullPageDemo.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-import { FullPageScatterplot } from "@haniffalab/cherita-react";
-
-export function FullPageDemo(props) {
-  return <FullPageScatterplot {...props} />;
-}
