--- sites/demo/src/containers/PseudospatialDemo.js
@@ -1,7 +1,0 @@
-import React from "react";
-
-import { FullPagePseudospatial } from "@haniffalab/cherita-react";
-
-export function PseudospatialDemo(props) {
-  return <FullPagePseudospatial {...props} />;
-}
