+++ src/scss/cherita.scss
@@ -58,7 +58,6 @@ $prefix: "bs-" !default;
 /* End scrollbar styling */
 
 .cherita-container-plot {
-  margin-top: 76px;
   padding: 20px;
   position: relative;
   min-height: 400px;
